+++ packages/snap-store-mobx/src/Autocomplete/Stores/AutocompleteTermStore.ts
@@ -2,14 +2,21 @@ import { observable, makeObservable } from 'mobx';
 import type { UrlManager } from '@searchspring/snap-url-manager';
 import type { AutocompleteStoreConfig, StoreServices } from '../../types';
 import { AutocompleteStateStore } from './AutocompleteStateStore';
-import { AutocompleteResponseModel } from '@athoscommerce/snapi-types';
+import { AutocompleteResponseModel, AutocompleteRequestModelSearchSourceEnum } from '@athoscommerce/snapi-types';
 
-type AutocompleteTermStoreConfig = Omit<TermData, 'data'> & {
+type AutocompleteTermStoreConfig = Omit<TermData, 'data' | 'type'> & {
 	config: AutocompleteStoreConfig;
 	data: {
 		autocomplete: AutocompleteResponseModel;
 	};
+	functions: {
+		resetTerms: () => void;
+	};
+	state: {
+		rootState: AutocompleteStateStore;
+	};
 };
+
 export class AutocompleteTermStore extends Array<Term> {
 	static get [Symbol.species](): ArrayConstructor {
 		return Array;
@@ -53,6 +60,9 @@ export class AutocompleteTermStore extends Array<Term> {
 						},
 						terms,
 					},
+					functions: params.functions,
+					state: params.state,
+					type: 'suggested' as AutocompleteRequestModelSearchSourceEnum,
 				})
 			)
 		);
@@ -67,7 +77,7 @@ export type TermData = {
 		resetTerms: () => void;
 	};
 	state: {
-		autocomplete: AutocompleteStateStore;
+		rootState: AutocompleteStateStore;
 	};
 	data: {
 		term: {
@@ -76,6 +86,7 @@ export type TermData = {
 		};
 		terms: Term[];
 	};
+	type: AutocompleteRequestModelSearchSourceEnum;
 };
 
 export class Term {
@@ -83,13 +94,15 @@ export class Term {
 	public value: string;
 	public preview: () => void;
 	public url: UrlManager;
+	public type: AutocompleteRequestModelSearchSourceEnum;
 
 	constructor(params: TermData) {
-		const { services, functions, state, data } = params || {};
+		const { services, functions, state, data, type } = params || {};
 		const { term, terms } = data || {};
 
 		this.active = term?.active;
 		this.value = term?.value;
+		this.type = type;
 
 		this.url = services?.urlManager?.set({ query: this.value });
 
@@ -98,10 +111,10 @@ export class Term {
 			terms.map((term) => {
 				term.active = false;
 			});
-
+			state.rootState.source = type;
 			this.active = true;
-			state.autocomplete.locks.terms.lock();
-			state.autocomplete.locks.facets.unlock();
+			state.rootState.locks.terms.lock();
+			state.rootState.locks.facets.unlock();
 
 			this.url?.set({ query: this.value }).go();
 		};