+++ packages/snap-store-mobx/src/types.ts
@@ -139,6 +139,8 @@ export type AutocompleteStoreConfigSettings = {
 		enabled?: boolean;
 		limit?: number;
 		showResults?: boolean;
+		max?: number;
+		url?: string;
 	};
 	redirects?: {
 		merchandising?: boolean;